chore(DeploymentServerUI/Graphs): Move Graphs to own folder
