+++ packages/deployment-server-ui/src/components/Graphs/DeploymentsByTypeGraph.tsx
